[af-263] refactor: dto package 추가
